et0 should be a floating point value, not int

+++ cimis.h
@@ -6,7 +6,7 @@ struct CIMIS_data{
 	int station;
 	char date[15];
 	int hour; // NOTE: this is in PST (not to be confused with PDT)
-	int et0;
+	double et0;
 };
 
 // gets the lastest information from the CIMIS FTP server and puts necessary information into the data struct
